pyproject.toml
[project]
name = "12-sport-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastexcel>=0.14.0",
    "ipykernel>=6.29.5",
    "kafka-python>=2.2.10",
    "pandas>=2.3.0",
    "polars>=1.30.0",
    "prefect>=3.4.9",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
]
